requirements.txt
# Web Frameworks
streamlit==1.29.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP & API
requests==2.31.0
httpx==0.25.1

# AI & LLM
openai>=1.12.0,<2.0.0
langchain>=0.1.20,<1.0.0
langchain-openai>=0.1.0,<1.0.0

# Data & Config
python-dotenv==1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Code Execution
httpx==0.25.1
aiofiles==23.2.1

# WebSocket Support (optional - for future enhancements)
# redis==5.0.1
# python-socketio==5.10.0
# aioredis==2.0.1